+++ servers/mcp-neo4j-data-modeling/src/mcp_neo4j_data_modeling/server.py
@@ -328,6 +328,26 @@ def create_mcp_server(namespace: str = "") -> FastMCP:
             "total_examples": len(examples),
             "usage": "Use the get_example_data_model tool with any of the example names above to get a specific data model",
         }
+    
+    @mcp.tool(name=namespace_prefix + "load_from_owl_turtle")
+    def load_from_owl_turtle(owl_turtle_str: str) -> DataModel:
+        """
+        Load a data model from an OWL Turtle string. 
+        This process is lossy and some components of the ontology may be lost in the data model schema.
+        Returns a DataModel object.
+        """
+        logger.info("Loading a data model from an OWL Turtle string.")
+        return DataModel.from_owl_turtle_str(owl_turtle_str)
+
+    @mcp.tool(name=namespace_prefix + "export_to_owl_turtle")
+    def export_to_owl_turtle(data_model: DataModel) -> str:
+        """
+        Export a data model to an OWL Turtle string. 
+        This process is lossy since OWL does not support properties on relationships.
+        Returns a string representation of the data model in OWL Turtle format.
+        """
+        logger.info("Exporting a data model to an OWL Turtle string.")
+        return data_model.to_owl_turtle_str()
 
     @mcp.prompt(title="Create New Data Model")
     def create_new_data_model(