+++ servers/mcp-neo4j-data-modeling/CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Changed
 
 ### Added
+* Add import and export methods to `DataModel` for turtle OWL strings
+* Add MCP tools for loading and exporting turtle OWL files
 
 ## v0.5.1